requirements.txt
flask
gunicorn
itsdangerous 
jinja2
python-dotenv
genai
